CV_MIND-main/CV_MIND-main/hf_space/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# Text extraction
PyMuPDF==1.24.5
python-docx==1.1.2

# ML
xgboost==2.1.1
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==1.26.4
